Trying to make iso script more robust.

+++ cmake-iso.sh
@@ -1,12 +1,16 @@
 #!/usr/bin/env sh
-
+GRUB_SRC="../include/grub.cfg"
 if grub-file --is-x86-multiboot ArtOS.bin; then
-  echo multiboot confirmed
-  mkdir -p isodir/boot/grub
-  cp ArtOS.bin isodir/boot/ArtOS.bin
-  cp ../include/grub.cfg isodir/boot/grub/grub.cfg
-  grub-mkrescue -o ArtOS.iso isodir
-  rm -rf isodir
+  if test -f $GRUB_SRC; then
+    echo multiboot confirmed
+    mkdir -p isodir/boot/grub
+    cp ArtOS.bin isodir/boot/ArtOS.bin
+    cp $GRUB_SRC isodir/boot/grub/grub.cfg
+    grub-mkrescue -o ArtOS.iso isodir
+    rm -rf isodir
+  else
+    echo could not find grub config
+  fi
 else
   echo the file is not multiboot
 fi
